Changed titleselector to domainselector
